Use custom model instead of wp.api to enable author id
I'm waiting for word back from the REST API team about a possible
issue with the author attribute. The post model has an author
attribute that is converted into a user object on construct, but
comment authors (when converted) have bogus user objects and get
malformed. When that happens, we lose the author id, preventing
us from knowing who the comment author was. Hoping to either
submit a PR or to help reproduce based on feedback.