First pass of optimistic/inline comment replies
Use Backbone to create a new CommentModel based on which comment
we're replying to and the contents of our form. Add to the
collection on click. Not syncing with the server yet.

Also switched from this.model.attributes to this.model in the
template as Backbone will throw an exception if author doesn't
exist but will proceed if attributes.author does not.

Still need to figure out the best approach for avatar and other
data, but I suspect adding as data attributes as done so far will
give enough data to build this out.