Add pre_insert_comment to verify data instead of trusting client
A bunch of work here to get only what we need from the client to
add comments. We can't trust anything they say implicitly, so do
the verification we're able to do using the REST-API filter.
For now, we're only checking that the current logged in user is
trying to add a comment by themselves. Support for anonymous
users (if enabled) and potentially other modes will follow.

Also added 56 to the list of avatar URL sizes to match the
twenty* themes, and added a user avatar link to data for new
comment display.