Better async processing for replies and edits
Needs some refactoring. This edit removes the optimistic insertion
of comments, so we'll need to put that back in. The issue was that
we still had a template that assumed we knew the id; since that's
not true until we hear back from the server, we'll need to put in
a placeholder id (or different template with hidden edit/reply
links, or something along those lines).

There's also some debug spam for testing.