Prepare comments before sending, with dates and edit status
We shouldn't necessarily be making our editing assumptions in
the JavaScript code. Use what we know from the server side to
tell us if we are able to edit a comment or not.

This also moves the date formatting to the server as well,
which is much more akin to the previous format. Unfortunately,
get_comment_time relies on the global $comment, and since we're
not in The Loop, we have to write our own time formatter.
Stick as close to the original functionality as possible.