Remove temp form, append to #wp-comment-content
In retrospect, we don't need the temporary form at all. This is
a holdout from the old template structure. We always return the
comment form div to the same location, so adding and removing a
placeholder div isn't necessary--we already know where it's
going to go.

Cut out the whole setupTempForm function. This eliminates the
need for a cache peek, so it's removed as well.