Don't empty element on view render, just as needed
We don't need to redraw the entire comment field each time.
Instead, draw comment items that haven't been rendered yet,
and empty the element when we do something like change the
current page.

Also, fix a couple of selector caches that shouldn't be
cached--these are dynamic selectors and might be added later.