Add two more translated strings
Had a wonderful discussion with my wife who speaks Québécois
french. The big conclusion, I think, is whether or not I used
the right terminology for "Alternate 1" and "Alternate 2".
The word alternate depends on context--is an alternate a person
or a thing, and if a thing, what type of thing? Is it masculine
or feminine? Fascinating. Translation is hard!